Refactoring of transactions modules specs
